Move err to utils package
