+++ packages/utils/src/index.ts
@@ -2,6 +2,7 @@ export * from "./logger/index.js";
 export * from "./yaml/index.js";
 export * from "./assert.js";
 export * from "./bytes.js";
+export * from "./err.js";
 export * from "./errors.js";
 export * from "./format.js";
 export * from "./map.js";
